SRIOV hotplug, e2e: add SRIOV hotplug test
